Addressing more PR comments.
