+++ src/Microsoft.ML.StandardLearners/Standard/MultiClass/Pkpd.cs
@@ -73,8 +73,6 @@ namespace Microsoft.ML.Runtime.Learners
         /// Developers should instantiate <see cref="Pkpd"/> by supplying the trainer argument directly to the <see cref="Pkpd"/> constructor
         /// using the other public constructor.
         /// </summary>
-        /// <param name="env"></param>
-        /// <param name="args"></param>
         public Pkpd(IHostEnvironment env, Arguments args)
             : base(env, args, LoadNameValue)
         {
@@ -85,7 +83,7 @@ namespace Microsoft.ML.Runtime.Learners
         /// </summary>
         /// <param name="env">The <see cref="IHostEnvironment"/> instance.</param>
         /// <param name="binaryEstimator">An instance of a binary <see cref="ITrainerEstimator{TTransformer, TPredictor}"/> used as the base trainer.</param>
-        /// <param name="calibrator">The <see cref="ICalibratorTrainer"/> used.</param>
+        /// <param name="calibrator">The calibrator. If a calibrator is not explicitely provided, it will default to <see cref="PlattCalibratorCalibratorTrainer"/></param>
         /// <param name="labelColumn">The name of the label colum.</param>
         /// <param name="imputeMissingLabelsAsNegative">Whether to treat missing labels as having negative labels, instead of keeping them missing.</param>
         /// <param name="maxCalibrationExamples">Number of instances to train the calibrator.</param>
