+++ test/Microsoft.ML.Tests/Scenarios/Api/Estimators/Metacomponents.cs
@@ -35,105 +35,5 @@ namespace Microsoft.ML.Tests.Scenarios.Api
                 var model = pipeline.Fit(data);
             }
         }
-
-        /// <summary>
-        /// OVA with calibrator
-        /// </summary>
-        [Fact]
-        public void New_OVAWithCalibrator()
-        {
-            var dataPath = GetDataPath(IrisDataPath);
-
-            using (var env = new TlcEnvironment())
-            {
-                var calibrator = new PavCalibratorTrainer(env);
-
-                var data = new TextLoader(env, MakeIrisTextLoaderArgs())
-                    .Read(new MultiFileSource(dataPath));
-
-                var sdcaTrainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments { MaxIterations = 100, Shuffle = true, NumThreads = 1 }, "Features", "Label");
-                var pipeline = new MyConcatTransform(env, "Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth")
-                    .Append(new TermEstimator(env, "Label"), TransformerScope.TrainTest)
-                    .Append(new Ova(env, sdcaTrainer, "Label", calibrator: calibrator, maxCalibrationExamples: 990000))
-                    .Append(new KeyToValueEstimator(env, "PredictedLabel"));
-
-                var model = pipeline.Fit(data);
-            }
-        }
-
-        /// <summary>
-        /// OVA with calibrator
-        /// </summary>
-        [Fact]
-        public void New_OVAWithAllConstructorArgs()
-        {
-            var dataPath = GetDataPath(IrisDataPath);
-
-            using (var env = new TlcEnvironment())
-            {
-                var calibrator = new PavCalibratorTrainer(env);
-
-                var data = new TextLoader(env, MakeIrisTextLoaderArgs())
-                    .Read(new MultiFileSource(dataPath));
-
-                var sdcaTrainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments { MaxIterations = 100, Shuffle = true, NumThreads = 1, Calibrator = null }, "Features", "Label");
-                var pipeline = new MyConcatTransform(env, "Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth")
-                    .Append(new TermEstimator(env, "Label"), TransformerScope.TrainTest)
-                    .Append(new Ova(env, sdcaTrainer, "Label", true, calibrator: calibrator, 10000, true))
-                    .Append(new KeyToValueEstimator(env, "PredictedLabel"));
-
-                var model = pipeline.Fit(data);
-            }
-        }
-
-        /// <summary>
-        /// OVA with uncalibrated
-        /// </summary>
-        [Fact]
-        public void New_OVAUncalibrated()
-        {
-            var dataPath = GetDataPath(IrisDataPath);
-
-            using (var env = new TlcEnvironment())
-            {
-                var calibrator = new PavCalibratorTrainer(env);
-
-                var data = new TextLoader(env, MakeIrisTextLoaderArgs())
-                    .Read(new MultiFileSource(dataPath));
-
-                var sdcaTrainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments { MaxIterations = 100, Shuffle = true, NumThreads = 1 }, "Features", "Label");
-                var pipeline = new MyConcatTransform(env, "Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth")
-                    .Append(new TermEstimator(env, "Label"), TransformerScope.TrainTest)
-                    .Append(new Ova(env, sdcaTrainer, useProbabilities: false))
-                    .Append(new KeyToValueEstimator(env, "PredictedLabel"));
-
-                var model = pipeline.Fit(data);
-            }
-        }
-
-        /// <summary>
-        /// OVA with calibrator
-        /// </summary>
-        [Fact]
-        public void New_Pkpd()
-        {
-            var dataPath = GetDataPath(IrisDataPath);
-
-            using (var env = new TlcEnvironment())
-            {
-                var calibrator = new PavCalibratorTrainer(env);
-
-                var data = new TextLoader(env, MakeIrisTextLoaderArgs())
-                    .Read(new MultiFileSource(dataPath));
-
-                var sdcaTrainer = new LinearClassificationTrainer(env, new LinearClassificationTrainer.Arguments { MaxIterations = 100, Shuffle = true, NumThreads = 1 }, "Features", "Label");
-                var pipeline = new MyConcatTransform(env, "Features", "SepalLength", "SepalWidth", "PetalLength", "PetalWidth")
-                    .Append(new TermEstimator(env, "Label"), TransformerScope.TrainTest)
-                    .Append(new Pkpd(env, sdcaTrainer))
-                    .Append(new KeyToValueEstimator(env, "PredictedLabel"));
-
-                var model = pipeline.Fit(data);
-            }
-        }
     }
 }
